LC-221 - Move static/functions from main to different file P1
