+++ main.py
@@ -6,63 +6,17 @@ import imp
 import Queue
 import messaging
 import gui
-import sys
 import logging
 import logging.config
-import requests
 import semantic_version
-import locale
 from collections import OrderedDict
 from modules.helper.parser import load_from_config_file
-from modules.helper.system import load_translations_keys
+from modules.helper.system import load_translations_keys, PYTHON_FOLDER, CONF_FOLDER, MAIN_CONF_FILE, MODULE_FOLDER, \
+    LOG_FOLDER, GUI_TAG, TRANSLATION_FOLDER, LOG_FILE, LOG_FORMAT, get_language, get_update
 from modules.helper.module import BaseModule
 
 VERSION = '0.3.5'
 SEM_VERSION = semantic_version.Version(VERSION)
-if hasattr(sys, 'frozen'):
-    PYTHON_FOLDER = os.path.dirname(sys.executable)
-else:
-    PYTHON_FOLDER = os.path.dirname(os.path.abspath('__file__'))
-TRANSLATION_FOLDER = os.path.join(PYTHON_FOLDER, "translations")
-CONF_FOLDER = os.path.join(PYTHON_FOLDER, "conf")
-MODULE_FOLDER = os.path.join(PYTHON_FOLDER, "modules")
-MAIN_CONF_FILE = os.path.join(CONF_FOLDER, "config.cfg")
-GUI_TAG = 'gui'
-
-LOG_FOLDER = os.path.join(PYTHON_FOLDER, "logs")
-if not os.path.exists(LOG_FOLDER):
-    os.makedirs(LOG_FOLDER)
-LOG_FILE = os.path.join(LOG_FOLDER, 'chat_log.log')
-LOG_FORMAT = logging.Formatter("%(asctime)s [%(name)s] [%(levelname)s]  %(message)s")
-
-LANGUAGE_DICT = {
-    'en_US': 'en',
-    'en_GB': 'en',
-    'ru_RU': 'ru'
-}
-
-
-def get_update():
-    github_url = "https://api.github.com/repos/DeForce/LalkaChat/releases"
-    try:
-        update_json = requests.get(github_url, timeout=1)
-        if update_json.status_code == 200:
-            update = False
-            update_url = None
-            update_list = update_json.json()
-            for update_item in update_list:
-                if semantic_version.Version.coerce(update_item['tag_name'].lstrip('v')) > SEM_VERSION:
-                    update = True
-                    update_url = update_item['html_url']
-            return update, update_url
-    except Exception as exc:
-        log.info("Got exception: {0}".format(exc))
-    return False, None
-
-
-def get_language():
-    local_name, local_encoding = locale.getdefaultlocale()
-    return LANGUAGE_DICT.get(local_name, 'en')
 
 
 def init():
@@ -150,7 +104,7 @@ def init():
 
     # Checking updates
     log.info("Checking for updates")
-    loaded_modules['main']['update'], loaded_modules['main']['update_url'] = get_update()
+    loaded_modules['main']['update'], loaded_modules['main']['update_url'] = get_update(SEM_VERSION)
     if loaded_modules['main']['update']:
         log.info("There is new update, please update!")
